+++ packages/adminui-vue/components/controls/USelectMultiple.vue
@@ -526,6 +526,18 @@ export default {
       this.$emit('input', filterFixed)
     },
 
+    /**
+     * May be used in the parent component via its $refs property in case
+     * options' reset is needed.
+     */
+    clearOptions () {
+      this.options.splice(0, this.options.length)
+
+      this.pageNum = 0
+      this.moreVisible = false
+      this.selectedOption = null
+    },
+
     isOptionClosable (option) {
       return !(this.readonly || !!this.isOptionFixed(option))
     },