+++ packages/ub/modules/ubErrors.js
@@ -4,21 +4,23 @@
  */
 
 /**
+ * @classdesc
  * Server-side Abort exception. To be used in server-side logic in case of HANDLED
  * exception. This errors logged using "Error" log level to prevent unnecessary
  * EXC log entries.
- *
- *       // UB client will show message inside <<<>>> to user (and translate it using UB.i18n)
- *       const UB = require('@unitybase/ub')
- *       throw new UB.UBAbort('<<<textToDisplayForClient>>>')
- *       // In case, client-side message shall be formatted:
- *       throw new UB.UBAbort('<<<file_not_found>>>', 'bad_file.name')
- *       // The "file_not_found" i18n string on client should be like `'File "{0}" is not found or not accessible'
- *       // Format args can be translated by assing a :i18n modifier to template string: `'File "{0:i18n}" is not found or not accessible'
- *
- *       // In case message should not be shown to the end used by ub-pub globalExceptionHandler `<<<>>>` can be omitted
- *       throw new UB.UBAbort('wrongParameters')
- *
+ * @example
+
+// UB client will show message inside <<<>>> to user (and translate it using UB.i18n)
+const UB = require('@unitybase/ub')
+throw new UB.UBAbort('<<<textToDisplayForClient>>>')
+// In case, client-side message shall be formatted:
+throw new UB.UBAbort('<<<file_not_found>>>', 'bad_file.name')
+// The "file_not_found" i18n string on client should be like `'File "{0}" is not found or not accessible'
+// Format args can be translated by assing a :i18n modifier to template string: `'File "{0:i18n}" is not found or not accessible'
+
+// In case message should not be shown to the end used by ub-pub globalExceptionHandler `<<<>>>` can be omitted
+throw new UB.UBAbort('wrongParameters')
+
  * @param {String} [message] Message
  * @extends {Error}
  * @constructor