+++ packages/adminui-vue/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `SignatureVerificationResult` - added a media type information (file/hardware)
 
 ### Changed