+++ packages/adminui-vue/CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UNavbarNotificationsButton`: added opening of unread notifications with the `messageType` === `system`
+  in the modal window for the user when logging into the system.
  - `u-table-many`: it's alternative for `u-select-many` component.
-    This component can be helpful when labels are very large or if you want to use table for working with
-    `"dataType": "Many"` entity attributes. 
-    See examples in file UTableMany.md.
+  This component can be helpful when labels are very large or if you want to use table for working with
+  `"dataType": "Many"` entity attributes. 
+  See examples in file UTableMany.md.
 ### Changed
 
 ### Deprecated