+++ packages/adminui-vue/CHANGELOG.md
@@ -6,14 +6,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- `UNavbarNotificationsButton`: added opening of unread notifications with the `messageType` === `system`
-  in the modal window for the user when logging into the system.
 
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+## [5.20.15] - 2021-06-14
+### Added
+ - `SignatureVerificationResult` - added a media type information (file/hardware)
+ - `window.capiSelectionDialog` - a Vue based interface for cryptographic module selection (used by UB EE/DE)
+ - `UNavbarNotificationsButton`: added opening of unread notifications with the `messageType` === `system`
+  in the modal window for the user when logging into the system.
+ - `processing.saved`: pass method name `insert/update` as second argument into a `saved` hook on form
+
+### Changed
+ - *BREAKING* UCodeMirror attribute renamed `readOnly` -> `readonly`;
+   This fix u-auto-field readonly binding in case attribute is on Json type 
 
 ### Fixed
 - `computedVuex` - pass `key` property to `SET` mutation